Added levels for all panels
