+++ scripts/addons/cam/ui.py
@@ -47,6 +47,9 @@ from cam.ui_panels.movement       import *
 from cam.ui_panels.feedrate       import *
 from cam.ui_panels.optimisation   import *
 from cam.ui_panels.area           import *
+from cam.ui_panels.gcode          import *
+from cam.ui_panels.pack           import *
+from cam.ui_panels.slice          import *
 
 
 class CAM_UL_orientations(UIList):
@@ -59,89 +62,6 @@ class CAM_UL_orientations(UIList):
             layout.label(text="", icon_value=icon)
 
 
-
-class CAM_GCODE_Panel(CAMButtonsPanel, bpy.types.Panel):
-    """CAM operation g-code options panel"""
-    bl_label = "CAM g-code options "
-    bl_idname = "WORLD_PT_CAM_GCODE"
-
-    COMPAT_ENGINES = {'BLENDERCAM_RENDER'}
-
-    def draw(self, context):
-        layout = self.layout
-        scene = bpy.context.scene
-        if len(scene.cam_operations) == 0:
-            layout.label(text='Add operation first')
-        if len(scene.cam_operations) > 0:
-            ao = scene.cam_operations[scene.cam_active_operation]
-            if ao.valid:
-                layout.prop(ao, 'output_header')
-
-                if ao.output_header:
-                    layout.prop(ao, 'gcode_header')
-                layout.prop(ao, 'output_trailer')
-                if ao.output_trailer:
-                    layout.prop(ao, 'gcode_trailer')
-                layout.prop(ao, 'enable_dust')
-                if ao.enable_dust:
-                    layout.prop(ao, 'gcode_start_dust_cmd')
-                    layout.prop(ao, 'gcode_stop_dust_cmd')
-                layout.prop(ao, 'enable_hold')
-                if ao.enable_hold:
-                    layout.prop(ao, 'gcode_start_hold_cmd')
-                    layout.prop(ao, 'gcode_stop_hold_cmd')
-                layout.prop(ao, 'enable_mist')
-                if ao.enable_mist:
-                    layout.prop(ao, 'gcode_start_mist_cmd')
-                    layout.prop(ao, 'gcode_stop_mist_cmd')
-
-            else:
-                layout.label(text='Enable Show experimental features')
-                layout.label(text='in Blender CAM Addon preferences')
-
-
-class CAM_PACK_Panel(CAMButtonsPanel, bpy.types.Panel):
-    """CAM material panel"""
-    bl_label = "Pack curves on sheet"
-    bl_idname = "WORLD_PT_CAM_PACK"
-
-    COMPAT_ENGINES = {'BLENDERCAM_RENDER'}
-
-    def draw(self, context):
-        layout = self.layout
-        scene = bpy.context.scene
-        settings = scene.cam_pack
-        layout.label(text='warning - algorithm is slow.')
-        layout.label(text='only for curves now.')
-
-        layout.operator("object.cam_pack_objects")
-        layout.prop(settings, 'sheet_fill_direction')
-        layout.prop(settings, 'sheet_x')
-        layout.prop(settings, 'sheet_y')
-        layout.prop(settings, 'distance')
-        layout.prop(settings, 'tolerance')
-        layout.prop(settings, 'rotate')
-        if settings.rotate:
-            layout.prop(settings, 'rotate_angle')
-
-
-class CAM_SLICE_Panel(CAMButtonsPanel, bpy.types.Panel):
-    """CAM slicer panel"""
-    bl_label = "Slice model to plywood sheets"
-    bl_idname = "WORLD_PT_CAM_SLICE"
-
-    COMPAT_ENGINES = {'BLENDERCAM_RENDER'}
-
-    def draw(self, context):
-        layout = self.layout
-        scene = bpy.context.scene
-        settings = scene.cam_slice
-
-        layout.operator("object.cam_slice_objects")
-        layout.prop(settings, 'slice_distance')
-        layout.prop(settings, 'slice_above0')
-        layout.prop(settings, 'slice_3d')
-        layout.prop(settings, 'indexes')
 
 
 # panel containing all tools
