+++ scripts/addons/cam/ui_panels/slice.py
@@ -1,0 +1,24 @@
+
+import bpy
+from cam.ui_panels.buttons_panel import CAMButtonsPanel
+
+
+class CAM_SLICE_Panel(CAMButtonsPanel, bpy.types.Panel):
+    """CAM slicer panel"""
+    bl_label = "Slice model to plywood sheets"
+    bl_idname = "WORLD_PT_CAM_SLICE"
+    panel_interface_level = 2
+
+    COMPAT_ENGINES = {'BLENDERCAM_RENDER'}
+
+    def draw(self, context):
+        layout = self.layout
+        scene = bpy.context.scene
+        settings = scene.cam_slice
+
+        layout.operator("object.cam_slice_objects")
+        layout.prop(settings, 'slice_distance')
+        layout.prop(settings, 'slice_above0')
+        layout.prop(settings, 'slice_3d')
+        layout.prop(settings, 'indexes')
+
